// src/redux/actions/types/chart.ts
enum ChartTypes {
  ADD_DATA = 'Chart/ADD_DATA',
  CHANGE_ACTIVE = 'Chart/CHANGE_ACTIVE',
  TOGGLE_CHART_LOADER = 'Chart/TOGGLE_CHART_LOADER',
  SET_CHECKED_ITEMS = 'Chart/SET_CHECKED_ITEMS',
  TOGLE_MSA = 'Chart/TOGLE_MSA',
  TOGLE_BINDING_GRAPH = 'Chart/TOGLE_BINDING_GRAPH',
  CHANGE_GRANULARITY = 'Chart/CHANGE_GRANULARITY',
  CHANGE_STORE_BY_HISTORY = 'Chart/CHANGE_STORE_BY_HISTORY',
  RESTORE_GRAPH_SETTINGS = 'Chart/RESTORE_GRAPH_SETTINGS',
  SET_SCROLL_POSITION = 'Chart/SET_SCROLL_POSITION',
}
export default ChartTypes;